chore: prepare X4 Pro Sync edition for public use
Firmware
- No automatic full-screen refresh while reading on any reader: the page
  cadence counter is gone (XTC included); a HALF refresh happens only on a
  manual refresh. The Refresh Frequency setting is removed.
- EPUB link taps follow the touchscreen switch, like tap page turns.
- about is advertised in download_kinds.
- One firmware status text shared by Settings and the Bluetooth/Firmware page.
- Unused BleLink members and debug logging removed; comments trimmed.

Build
- src/network/BuildStamp.h is generated by scripts/build_stamp.py.
- scripts/make_firmware_json.sh publishes an image plus firmware.json for the
  X4 Pro Sync app's update page.
- The Uc8279X4 settle-before-power-off fix is carried as a freeink-sdk patch
  applied by scripts/patch_freeink_sdk.py.

Docs
- README, USER_GUIDE and AGENTS describe the X4 Pro edition; the BLE protocol
  doc covers heartbeat fields, about, delete_book, start_put version/replace
  and host_name. The obsolete BLE transfer design spec is removed.

Verified: pio run -e x4pro, -e x4c and -e papermono build. -e default fails in
the build image before compiling (SCons.Tool.FortranCommon missing).

+++ src/activities/reader/EpubReaderMenuActivity.h
@@ -38,7 +38,7 @@ class EpubReaderMenuActivity final : public UiListActivity {
   static void buildMenuItems(std::vector<MenuItem>& items, bool hasFootnotes, bool hasBookmarks);
 
   explicit EpubReaderMenuActivity(GfxRenderer& renderer, MappedInputManager& mappedInput, const std::string& title,
-                                  const int currentPage, const int totalPages, const int bookProgressPercent,
+                                  const int currentPage, const int totalPages, const float bookProgressPercent,
                                   const uint8_t currentOrientation, const bool hasFootnotes, bool hasBookmarks);
 
   void render(RenderLock&&) override;
@@ -78,5 +78,7 @@ class EpubReaderMenuActivity final : public UiListActivity {
   const std::vector<const char*> pageTurnLabels = {I18N.get(StrId::STR_STATE_OFF), "1", "3", "6", "12"};
   int currentPage = 0;
   int totalPages = 0;
-  int bookProgressPercent = 0;
+  // Float, so the menu and the reader's top bar cannot disagree about the same
+  // book. Rounding this to an int made them differ by up to half a percent.
+  float bookProgressPercent = 0.0f;
 };
